Reimplement get_local_elevation and tile_to_world

+++ src/zthabitatmgr.rs
@@ -42,13 +42,14 @@ impl ZTHabitatMgr {
         self.get_habitat(tile.pos.x, tile.pos.y)
     }
 
-    pub fn get_habitat(&self, pos_x: u32, pos_y: u32) -> Option<ZTHabitat> {
+    // TODO: Should return Option<ZTExhbit> where ZTExhibit is a enum of ZTHabitat or ZTTankExhibit
+    pub fn get_habitat(&self, pos_x: i32, pos_y: i32) -> Option<ZTHabitat> {
         let base_ptr = self.other_array_start;
-        let offset_1 = pos_x * 0xc;
+        let offset_1 = pos_x as u32 * 0xc;
         let intermediate_ptr = get_from_memory::<u32>(base_ptr + offset_1);
 
 
-        let offset_2 = pos_y * 0x28;
+        let offset_2 = pos_y as u32 * 0x28;
         let ptr = get_from_memory::<u32>(intermediate_ptr + offset_2);
 
         info!("Habitat ptr: {:#x}", ptr);
